Fix the URL for the package loading file
I've moved things around a bit since I wrote this blog post and, without
thinking at the time, I'd just linked to the file in master rather than at
that point in time. So I've swapped the URL that was there for one that's at
the last point in time where the file existed.