:sparkles: Place humans.txt in the root of the filesystem
Sadly it doesn't look like Pelican, or Flex, allow for the easy addition of
arbitrary headers in the head of the output HTML; so at the moment it looks
like I can't easily add the link for this. But let's get the file there
anyway just in case someone looks for it.